Modularize protocol envelope definitions and update line limits
