--- crates/protocol/src/envelope.rs
@@ -1,1392 +1,0 @@
-use core::convert::TryFrom;
-use core::fmt;
-use core::str::FromStr;
-
-use std::string::String;
-
-/// Number of bytes in a multiplexed rsync message header.
-pub const HEADER_LEN: usize = 4;
-
-/// Maximum payload length representable in a multiplexed header.
-pub const MAX_PAYLOAD_LENGTH: u32 = 0x00FF_FFFF;
-
-/// Base offset added to multiplexed message codes when encoding headers.
-///
-/// Upstream rsync defines `MPLEX_BASE` as the separation point between raw data
-/// and control messages flowing over the multiplexed stream. Tags transmitted on
-/// the wire add this offset to the numeric [`MessageCode`] value so the high
-/// header byte can be inspected quickly. Exposing the constant keeps the Rust
-/// implementation in sync with the C reference and avoids duplicating the magic
-/// value across crates that need to reason about multiplexed tags.
-pub const MPLEX_BASE: u8 = 7;
-const PAYLOAD_MASK: u32 = 0x00FF_FFFF;
-
-/// Log classification used by upstream rsync's `enum logcode` table.
-///
-/// The numeric values mirror the identifiers found in `rsync.h` so the logging
-/// subsystem can translate between multiplexed tags and the log severities used
-/// by upstream traces. While only a subset of log codes flow over the
-/// multiplexed stream, the complete enum is provided for parity (including
-/// `FNONE`, which upstream reserves for internal use).
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[repr(u8)]
-pub enum LogCode {
-    #[doc(alias = "FNONE")]
-    /// Placeholder that is never transmitted on the wire (`FNONE`).
-    None = 0,
-    #[doc(alias = "FERROR_XFER")]
-    /// Fatal transfer error (`FERROR_XFER`).
-    ErrorXfer = 1,
-    #[doc(alias = "FINFO")]
-    /// Informational log message (`FINFO`).
-    Info = 2,
-    #[doc(alias = "FERROR")]
-    /// Non-fatal error (`FERROR`).
-    Error = 3,
-    #[doc(alias = "FWARNING")]
-    /// Warning message (`FWARNING`).
-    Warning = 4,
-    #[doc(alias = "FERROR_SOCKET")]
-    /// Error emitted by the sibling process over the receiver/generator pipe
-    /// (`FERROR_SOCKET`).
-    ErrorSocket = 5,
-    #[doc(alias = "FLOG")]
-    /// Log message only written to the daemon logs (`FLOG`).
-    Log = 6,
-    #[doc(alias = "FCLIENT")]
-    /// Client-only message (`FCLIENT`).
-    Client = 7,
-    #[doc(alias = "FERROR_UTF8")]
-    /// UTF-8 conversion problem reported by a sibling (`FERROR_UTF8`).
-    ErrorUtf8 = 8,
-}
-
-impl LogCode {
-    /// Ordered list of all log codes understood by rsync 3.4.1.
-    pub const ALL: [LogCode; 9] = [
-        LogCode::None,
-        LogCode::ErrorXfer,
-        LogCode::Info,
-        LogCode::Error,
-        LogCode::Warning,
-        LogCode::ErrorSocket,
-        LogCode::Log,
-        LogCode::Client,
-        LogCode::ErrorUtf8,
-    ];
-
-    /// Returns the ordered list of all log codes.
-    #[must_use]
-    pub const fn all() -> &'static [LogCode; 9] {
-        &Self::ALL
-    }
-
-    /// Returns the numeric representation expected on the wire.
-    #[must_use]
-    #[inline]
-    pub const fn as_u8(self) -> u8 {
-        self as u8
-    }
-
-    /// Attempts to construct a [`LogCode`] from its numeric representation.
-    #[must_use]
-    pub const fn from_u8(value: u8) -> Option<Self> {
-        match value {
-            0 => Some(Self::None),
-            1 => Some(Self::ErrorXfer),
-            2 => Some(Self::Info),
-            3 => Some(Self::Error),
-            4 => Some(Self::Warning),
-            5 => Some(Self::ErrorSocket),
-            6 => Some(Self::Log),
-            7 => Some(Self::Client),
-            8 => Some(Self::ErrorUtf8),
-            _ => None,
-        }
-    }
-
-    /// Returns the upstream `F*` identifier associated with this log code.
-    #[must_use]
-    pub const fn name(self) -> &'static str {
-        match self {
-            LogCode::None => "FNONE",
-            LogCode::ErrorXfer => "FERROR_XFER",
-            LogCode::Info => "FINFO",
-            LogCode::Error => "FERROR",
-            LogCode::Warning => "FWARNING",
-            LogCode::ErrorSocket => "FERROR_SOCKET",
-            LogCode::Log => "FLOG",
-            LogCode::Client => "FCLIENT",
-            LogCode::ErrorUtf8 => "FERROR_UTF8",
-        }
-    }
-}
-
-impl fmt::Display for LogCode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.name())
-    }
-}
-
-impl TryFrom<u8> for LogCode {
-    type Error = ParseLogCodeError;
-
-    fn try_from(value: u8) -> Result<Self, ParseLogCodeError> {
-        Self::from_u8(value).ok_or(ParseLogCodeError::new(value))
-    }
-}
-
-impl From<LogCode> for u8 {
-    fn from(value: LogCode) -> Self {
-        value.as_u8()
-    }
-}
-
-impl FromStr for LogCode {
-    type Err = ParseLogCodeError;
-
-    fn from_str(name: &str) -> Result<Self, Self::Err> {
-        match name {
-            "FNONE" => Ok(Self::None),
-            "FERROR_XFER" => Ok(Self::ErrorXfer),
-            "FINFO" => Ok(Self::Info),
-            "FERROR" => Ok(Self::Error),
-            "FWARNING" => Ok(Self::Warning),
-            "FERROR_SOCKET" => Ok(Self::ErrorSocket),
-            "FLOG" => Ok(Self::Log),
-            "FCLIENT" => Ok(Self::Client),
-            "FERROR_UTF8" => Ok(Self::ErrorUtf8),
-            other => Err(ParseLogCodeError::new_name(other)),
-        }
-    }
-}
-
-/// Error returned when parsing a log code from its representation fails.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseLogCodeError {
-    /// The provided numeric identifier is not known to rsync 3.4.1.
-    InvalidValue(u8),
-    /// The provided mnemonic name is not known to rsync 3.4.1.
-    InvalidName(String),
-}
-
-impl ParseLogCodeError {
-    /// Creates a parse error that records the invalid numeric value.
-    #[must_use]
-    pub const fn new(invalid_value: u8) -> Self {
-        Self::InvalidValue(invalid_value)
-    }
-
-    /// Creates a parse error that records the invalid mnemonic name.
-    #[must_use]
-    pub fn new_name(invalid_name: &str) -> Self {
-        Self::InvalidName(invalid_name.to_owned())
-    }
-
-    /// Returns the numeric value that failed to parse, when available.
-    #[must_use]
-    pub const fn invalid_value(&self) -> Option<u8> {
-        match self {
-            Self::InvalidValue(value) => Some(*value),
-            Self::InvalidName(_) => None,
-        }
-    }
-
-    /// Returns the mnemonic name that failed to parse, when available.
-    #[must_use]
-    pub fn invalid_name(&self) -> Option<&str> {
-        match self {
-            Self::InvalidValue(_) => None,
-            Self::InvalidName(name) => Some(name.as_str()),
-        }
-    }
-}
-
-impl fmt::Display for ParseLogCodeError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::InvalidValue(value) => write!(f, "unknown log code value: {value}"),
-            Self::InvalidName(name) => write!(f, "unknown log code name: \"{name}\""),
-        }
-    }
-}
-
-impl std::error::Error for ParseLogCodeError {}
-
-/// Tags used for multiplexed messages flowing over the rsync protocol stream.
-///
-/// The numeric values mirror the upstream `enum msgcode` definitions so that
-/// higher layers can reason about message semantics without translating between
-/// Rust and C identifiers. Values that alias upstream `enum logcode`
-/// definitions retain their historic numbering to ensure interoperability with
-/// existing daemons.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[repr(u8)]
-pub enum MessageCode {
-    #[doc(alias = "MSG_DATA")]
-    /// Raw file data written to the multiplexed stream.
-    Data = 0,
-    #[doc(alias = "MSG_ERROR_XFER")]
-    /// Fatal transfer error (`FERROR_XFER`).
-    ErrorXfer = 1,
-    #[doc(alias = "MSG_INFO")]
-    /// Informational log message (`FINFO`).
-    Info = 2,
-    #[doc(alias = "MSG_ERROR")]
-    /// Non-fatal error (`FERROR`).
-    Error = 3,
-    #[doc(alias = "MSG_WARNING")]
-    /// Warning message (`FWARNING`).
-    Warning = 4,
-    #[doc(alias = "MSG_ERROR_SOCKET")]
-    /// Error emitted by the sibling process over the receiver/generator pipe
-    /// (`FERROR_SOCKET`).
-    ErrorSocket = 5,
-    #[doc(alias = "MSG_LOG")]
-    /// Log message only written to the daemon logs (`FLOG`).
-    Log = 6,
-    #[doc(alias = "MSG_CLIENT")]
-    /// Client-only message (`FCLIENT`).
-    Client = 7,
-    #[doc(alias = "MSG_ERROR_UTF8")]
-    /// UTF-8 conversion problem reported by a sibling (`FERROR_UTF8`).
-    ErrorUtf8 = 8,
-    #[doc(alias = "MSG_REDO")]
-    /// Request to reprocess a specific file-list index.
-    Redo = 9,
-    #[doc(alias = "MSG_STATS")]
-    /// Transfer statistics destined for the generator.
-    Stats = 10,
-    #[doc(alias = "MSG_IO_ERROR")]
-    /// Sender encountered an I/O error while accessing the source tree.
-    IoError = 22,
-    #[doc(alias = "MSG_IO_TIMEOUT")]
-    /// Daemon communicating its timeout to the peer.
-    IoTimeout = 33,
-    #[doc(alias = "MSG_NOOP")]
-    /// Legacy no-op message (protocol 30 compatibility).
-    NoOp = 42,
-    #[doc(alias = "MSG_ERROR_EXIT")]
-    /// Synchronizes an error exit across processes (protocol ≥ 31).
-    ErrorExit = 86,
-    #[doc(alias = "MSG_SUCCESS")]
-    /// Receiver reports a successfully updated file.
-    Success = 100,
-    #[doc(alias = "MSG_DELETED")]
-    /// Receiver reports a deleted file.
-    Deleted = 101,
-    #[doc(alias = "MSG_NO_SEND")]
-    /// Sender failed to open a requested file.
-    NoSend = 102,
-}
-
-/// Error returned when parsing a multiplexed message code from its mnemonic name fails.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ParseMessageCodeError {
-    invalid_name: String,
-}
-
-impl ParseMessageCodeError {
-    /// Creates a parse error that records the invalid mnemonic name.
-    #[must_use]
-    pub fn new(invalid_name: &str) -> Self {
-        Self {
-            invalid_name: invalid_name.to_owned(),
-        }
-    }
-
-    /// Returns the mnemonic name that failed to parse.
-    #[must_use]
-    pub fn invalid_name(&self) -> &str {
-        &self.invalid_name
-    }
-}
-
-impl fmt::Display for ParseMessageCodeError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "unknown multiplexed message code name: \"{}\"",
-            self.invalid_name
-        )
-    }
-}
-
-impl std::error::Error for ParseMessageCodeError {}
-
-/// Errors that arise when converting between [`LogCode`] and [`MessageCode`].
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum LogCodeConversionError {
-    /// The [`LogCode`] has no multiplexed [`MessageCode`] equivalent.
-    NoMessageEquivalent(LogCode),
-    /// The [`MessageCode`] does not map to a [`LogCode`].
-    NoLogEquivalent(MessageCode),
-}
-
-impl LogCodeConversionError {
-    /// Returns the [`LogCode`] that could not be converted, when available.
-    #[must_use]
-    pub const fn log_code(self) -> Option<LogCode> {
-        match self {
-            Self::NoMessageEquivalent(log) => Some(log),
-            Self::NoLogEquivalent(_) => None,
-        }
-    }
-
-    /// Returns the [`MessageCode`] that could not be converted, when available.
-    #[must_use]
-    pub const fn message_code(self) -> Option<MessageCode> {
-        match self {
-            Self::NoMessageEquivalent(_) => None,
-            Self::NoLogEquivalent(code) => Some(code),
-        }
-    }
-}
-
-impl fmt::Display for LogCodeConversionError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::NoMessageEquivalent(log) => {
-                write!(f, "log code {log} has no multiplexed message equivalent",)
-            }
-            Self::NoLogEquivalent(code) => {
-                write!(f, "message code {code} has no log code equivalent")
-            }
-        }
-    }
-}
-
-impl std::error::Error for LogCodeConversionError {}
-
-impl TryFrom<LogCode> for MessageCode {
-    type Error = LogCodeConversionError;
-
-    fn try_from(value: LogCode) -> Result<Self, LogCodeConversionError> {
-        MessageCode::from_log_code(value).ok_or(LogCodeConversionError::NoMessageEquivalent(value))
-    }
-}
-
-impl MessageCode {
-    /// Alias constant representing the legacy `MSG_FLUSH` identifier.
-    ///
-    /// Upstream rsync exposes `MSG_FLUSH` as a preprocessor macro that maps to
-    /// the same numeric value as [`MessageCode::Info`]. Maintaining the alias
-    /// allows callers to reference the historic name when mirroring traces or
-    /// constructing golden streams while still reusing the canonical `Info`
-    /// variant for on-the-wire encoding.
-    pub const FLUSH: MessageCode = MessageCode::Info;
-
-    /// Returns the numeric representation expected on the wire.
-    #[must_use]
-    #[inline]
-    pub const fn as_u8(self) -> u8 {
-        self as u8
-    }
-
-    /// Attempts to construct a [`MessageCode`] from its on-the-wire numeric representation.
-    ///
-    /// The mapping mirrors the upstream `enum msgcode` table and can be used in const
-    /// contexts where [`TryFrom<u8>`] is not available. Invalid values yield `None`, allowing
-    /// callers to fall back to [`EnvelopeError::UnknownMessageCode`] for diagnostics.
-    #[must_use]
-    pub const fn from_u8(value: u8) -> Option<Self> {
-        match value {
-            0 => Some(Self::Data),
-            1 => Some(Self::ErrorXfer),
-            2 => Some(Self::Info),
-            3 => Some(Self::Error),
-            4 => Some(Self::Warning),
-            5 => Some(Self::ErrorSocket),
-            6 => Some(Self::Log),
-            7 => Some(Self::Client),
-            8 => Some(Self::ErrorUtf8),
-            9 => Some(Self::Redo),
-            10 => Some(Self::Stats),
-            22 => Some(Self::IoError),
-            33 => Some(Self::IoTimeout),
-            42 => Some(Self::NoOp),
-            86 => Some(Self::ErrorExit),
-            100 => Some(Self::Success),
-            101 => Some(Self::Deleted),
-            102 => Some(Self::NoSend),
-            _ => None,
-        }
-    }
-
-    /// Ordered list of all message codes understood by rsync 3.4.1.
-    ///
-    /// The variants are arranged by their numeric value so that callers can
-    /// iterate deterministically when constructing golden multiplexed streams
-    /// or exhaustively testing round-trips. The ordering mirrors upstream's
-    /// `enum msgcode` definitions to preserve byte-level parity.
-    pub const ALL: [MessageCode; 18] = [
-        MessageCode::Data,
-        MessageCode::ErrorXfer,
-        MessageCode::Info,
-        MessageCode::Error,
-        MessageCode::Warning,
-        MessageCode::ErrorSocket,
-        MessageCode::Log,
-        MessageCode::Client,
-        MessageCode::ErrorUtf8,
-        MessageCode::Redo,
-        MessageCode::Stats,
-        MessageCode::IoError,
-        MessageCode::IoTimeout,
-        MessageCode::NoOp,
-        MessageCode::ErrorExit,
-        MessageCode::Success,
-        MessageCode::Deleted,
-        MessageCode::NoSend,
-    ];
-
-    /// Returns the ordered list of all known message codes.
-    #[must_use]
-    pub const fn all() -> &'static [MessageCode; 18] {
-        &Self::ALL
-    }
-
-    /// Reports whether this message carries human-readable logging output.
-    ///
-    /// Upstream rsync routes multiplexed messages tagged with the `FINFO`,
-    /// `FERROR*`, `FWARNING`, `FLOG`, and `FCLIENT` log codes directly to the
-    /// logging subsystem inside `read_a_msg()` (see `io.c`). Messages in this
-    /// category contain UTF-8 text payloads destined for the user or daemon
-    /// logs rather than control data. The Rust implementation mirrors that
-    /// classification so higher layers can dispatch log records without
-    /// duplicating the tag match logic.
-    #[must_use]
-    pub const fn is_logging(self) -> bool {
-        self.log_code().is_some()
-    }
-
-    /// Returns the log code associated with this message code when the payload
-    /// represents logging output.
-    #[must_use]
-    pub const fn log_code(self) -> Option<LogCode> {
-        match self {
-            MessageCode::ErrorXfer => Some(LogCode::ErrorXfer),
-            MessageCode::Info => Some(LogCode::Info),
-            MessageCode::Error => Some(LogCode::Error),
-            MessageCode::Warning => Some(LogCode::Warning),
-            MessageCode::ErrorSocket => Some(LogCode::ErrorSocket),
-            MessageCode::Log => Some(LogCode::Log),
-            MessageCode::Client => Some(LogCode::Client),
-            MessageCode::ErrorUtf8 => Some(LogCode::ErrorUtf8),
-            _ => None,
-        }
-    }
-
-    /// Returns the multiplexed message code associated with a log code when a
-    /// one-to-one mapping exists.
-    ///
-    /// Upstream rsync defines the `MSG_*` logging variants with numeric values
-    /// that mirror the entries in `enum logcode`. The Rust implementation
-    /// exposes the inverse mapping so that higher layers can start from the
-    /// log classification (for example when translating user-specified
-    /// `--info` or `--debug` categories) and recover the multiplexed tag used
-    /// on the wire. Variants such as [`LogCode::None`] do not have a
-    /// corresponding multiplexed tag and therefore yield `None`.
-    #[must_use]
-    pub const fn from_log_code(log: LogCode) -> Option<Self> {
-        match log {
-            LogCode::ErrorXfer => Some(Self::ErrorXfer),
-            LogCode::Info => Some(Self::Info),
-            LogCode::Error => Some(Self::Error),
-            LogCode::Warning => Some(Self::Warning),
-            LogCode::ErrorSocket => Some(Self::ErrorSocket),
-            LogCode::Log => Some(Self::Log),
-            LogCode::Client => Some(Self::Client),
-            LogCode::ErrorUtf8 => Some(Self::ErrorUtf8),
-            LogCode::None => None,
-        }
-    }
-
-    /// Returns the upstream `MSG_*` identifier associated with this code.
-    ///
-    /// The mapping mirrors `enum msgcode` in upstream rsync's `rsync.h` so that
-    /// diagnostics and tracing can surface the same mnemonic names that C
-    /// implementations print through helpers such as `get_mplex_name()`. Keeping
-    /// the strings identical makes it easier to compare mixed Rust/C traces when
-    /// validating parity.
-    #[must_use]
-    pub const fn name(self) -> &'static str {
-        match self {
-            MessageCode::Data => "MSG_DATA",
-            MessageCode::ErrorXfer => "MSG_ERROR_XFER",
-            MessageCode::Info => "MSG_INFO",
-            MessageCode::Error => "MSG_ERROR",
-            MessageCode::Warning => "MSG_WARNING",
-            MessageCode::ErrorSocket => "MSG_ERROR_SOCKET",
-            MessageCode::Log => "MSG_LOG",
-            MessageCode::Client => "MSG_CLIENT",
-            MessageCode::ErrorUtf8 => "MSG_ERROR_UTF8",
-            MessageCode::Redo => "MSG_REDO",
-            MessageCode::Stats => "MSG_STATS",
-            MessageCode::IoError => "MSG_IO_ERROR",
-            MessageCode::IoTimeout => "MSG_IO_TIMEOUT",
-            MessageCode::NoOp => "MSG_NOOP",
-            MessageCode::ErrorExit => "MSG_ERROR_EXIT",
-            MessageCode::Success => "MSG_SUCCESS",
-            MessageCode::Deleted => "MSG_DELETED",
-            MessageCode::NoSend => "MSG_NO_SEND",
-        }
-    }
-}
-
-impl TryFrom<MessageCode> for LogCode {
-    type Error = LogCodeConversionError;
-
-    fn try_from(value: MessageCode) -> Result<Self, LogCodeConversionError> {
-        value
-            .log_code()
-            .ok_or(LogCodeConversionError::NoLogEquivalent(value))
-    }
-}
-
-impl fmt::Display for MessageCode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.name())
-    }
-}
-
-impl TryFrom<u8> for MessageCode {
-    type Error = EnvelopeError;
-
-    fn try_from(value: u8) -> Result<Self, EnvelopeError> {
-        Self::from_u8(value).ok_or(EnvelopeError::UnknownMessageCode(value))
-    }
-}
-
-impl FromStr for MessageCode {
-    type Err = ParseMessageCodeError;
-
-    fn from_str(name: &str) -> Result<Self, Self::Err> {
-        match name {
-            "MSG_DATA" => Ok(Self::Data),
-            "MSG_ERROR_XFER" => Ok(Self::ErrorXfer),
-            "MSG_INFO" => Ok(Self::Info),
-            "MSG_FLUSH" => Ok(Self::Info),
-            "MSG_ERROR" => Ok(Self::Error),
-            "MSG_WARNING" => Ok(Self::Warning),
-            "MSG_ERROR_SOCKET" => Ok(Self::ErrorSocket),
-            "MSG_LOG" => Ok(Self::Log),
-            "MSG_CLIENT" => Ok(Self::Client),
-            "MSG_ERROR_UTF8" => Ok(Self::ErrorUtf8),
-            "MSG_REDO" => Ok(Self::Redo),
-            "MSG_STATS" => Ok(Self::Stats),
-            "MSG_IO_ERROR" => Ok(Self::IoError),
-            "MSG_IO_TIMEOUT" => Ok(Self::IoTimeout),
-            "MSG_NOOP" => Ok(Self::NoOp),
-            "MSG_ERROR_EXIT" => Ok(Self::ErrorExit),
-            "MSG_SUCCESS" => Ok(Self::Success),
-            "MSG_DELETED" => Ok(Self::Deleted),
-            "MSG_NO_SEND" => Ok(Self::NoSend),
-            other => Err(ParseMessageCodeError::new(other)),
-        }
-    }
-}
-
-impl From<MessageCode> for u8 {
-    fn from(value: MessageCode) -> Self {
-        value.as_u8()
-    }
-}
-
-/// Failures encountered while parsing or constructing multiplexed message headers.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum EnvelopeError {
-    /// Fewer than [`HEADER_LEN`] bytes were provided when attempting to encode or decode a
-    /// header.
-    TruncatedHeader {
-        /// Number of bytes that were available when the operation began.
-        actual: usize,
-    },
-    /// The high tag byte did not include the required [`MPLEX_BASE`] offset.
-    InvalidTag(u8),
-    /// The encoded message code is not understood by rsync 3.4.1.
-    UnknownMessageCode(u8),
-    /// The payload length exceeded the representable range.
-    OversizedPayload(u32),
-}
-
-impl fmt::Display for EnvelopeError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::TruncatedHeader { actual } => {
-                write!(
-                    f,
-                    "multiplexed header truncated: expected {HEADER_LEN} bytes, got {actual}"
-                )
-            }
-            Self::InvalidTag(tag) => {
-                write!(f, "multiplexed header contained invalid tag byte {tag}")
-            }
-            Self::UnknownMessageCode(code) => {
-                write!(f, "unknown multiplexed message code {code}")
-            }
-            Self::OversizedPayload(len) => {
-                write!(
-                    f,
-                    "multiplexed payload length {len} exceeds maximum {MAX_PAYLOAD_LENGTH}"
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for EnvelopeError {}
-
-/// A fully decoded multiplexed message header.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct MessageHeader {
-    code: MessageCode,
-    payload_len: u32,
-}
-
-impl MessageHeader {
-    /// Creates a new header for `code` with the provided payload length.
-    ///
-    /// The constructor is `const` so golden multiplexed streams can be
-    /// assembled at compile time, matching upstream rsync's use of static
-    /// header tables for protocol tests.
-    pub const fn new(code: MessageCode, payload_len: u32) -> Result<Self, EnvelopeError> {
-        if payload_len > MAX_PAYLOAD_LENGTH {
-            return Err(EnvelopeError::OversizedPayload(payload_len));
-        }
-
-        Ok(Self { code, payload_len })
-    }
-
-    /// Parses a header from the beginning of `bytes`.
-    pub fn decode(bytes: &[u8]) -> Result<Self, EnvelopeError> {
-        if bytes.len() < HEADER_LEN {
-            return Err(EnvelopeError::TruncatedHeader {
-                actual: bytes.len(),
-            });
-        }
-
-        let mut encoded = [0u8; HEADER_LEN];
-        encoded.copy_from_slice(&bytes[..HEADER_LEN]);
-        let raw = u32::from_le_bytes(encoded);
-
-        Self::from_raw(raw)
-    }
-
-    /// Constructs a multiplexed header from the raw 32-bit representation used on the wire.
-    ///
-    /// The raw value is interpreted in host-endian order after decoding the little-endian bytes
-    /// emitted by upstream rsync. This mirrors the layout produced by [`MessageHeader::encode`] and
-    /// allows golden multiplexed streams to embed headers without materialising temporary byte
-    /// arrays.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rsync_protocol::{MessageCode, MessageHeader};
-    ///
-    /// let original = MessageHeader::new(MessageCode::Warning, 3).unwrap();
-    /// let raw = original.encode_raw();
-    /// let decoded = MessageHeader::from_raw(raw).unwrap();
-    ///
-    /// assert_eq!(decoded, original);
-    /// ```
-    #[must_use = "Discard the decoded header only after handling potential validation errors"]
-    #[inline]
-    pub const fn from_raw(raw: u32) -> Result<Self, EnvelopeError> {
-        let tag = (raw >> 24) as u8;
-        if tag < MPLEX_BASE {
-            return Err(EnvelopeError::InvalidTag(tag));
-        }
-
-        let code_value = tag - MPLEX_BASE;
-        match MessageCode::from_u8(code_value) {
-            Some(code) => Self::new(code, raw & PAYLOAD_MASK),
-            None => Err(EnvelopeError::UnknownMessageCode(code_value)),
-        }
-    }
-
-    /// Encodes this header into the little-endian format used on the wire.
-    #[must_use]
-    pub const fn encode(self) -> [u8; HEADER_LEN] {
-        self.encode_raw().to_le_bytes()
-    }
-
-    /// Encodes this header into the caller-provided slice without allocating.
-    ///
-    /// The helper mirrors upstream rsync's practice of writing directly into
-    /// stack-allocated buffers when constructing multiplexed messages. Callers
-    /// that already own suitably sized scratch space can avoid the temporary
-    /// array returned by [`MessageHeader::encode`] by reusing this method
-    /// instead. When the destination slice is smaller than [`HEADER_LEN`], an
-    /// [`EnvelopeError::TruncatedHeader`] error is returned and no bytes are
-    /// written.
-    pub fn encode_into_slice(self, out: &mut [u8]) -> Result<(), EnvelopeError> {
-        if out.len() < HEADER_LEN {
-            return Err(EnvelopeError::TruncatedHeader { actual: out.len() });
-        }
-
-        let raw = self.encode_raw();
-        out[..HEADER_LEN].copy_from_slice(&raw.to_le_bytes());
-        Ok(())
-    }
-
-    /// Returns the raw 32-bit representation used for little-endian multiplexed headers.
-    ///
-    /// Callers that need to embed headers in compile-time tables can use this helper to avoid
-    /// materialising intermediate `[u8; 4]` arrays. The returned value matches the integer obtained
-    /// by interpreting the encoded bytes in host-endian order.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rsync_protocol::{MessageCode, MessageHeader};
-    ///
-    /// let header = MessageHeader::new(MessageCode::Info, 7).unwrap();
-    /// let raw = header.encode_raw();
-    /// assert_eq!(MessageHeader::from_raw(raw).unwrap(), header);
-    /// ```
-    #[must_use]
-    #[inline]
-    pub const fn encode_raw(self) -> u32 {
-        let tag = (MPLEX_BASE as u32) + (self.code as u32);
-        (tag << 24) | (self.payload_len & PAYLOAD_MASK)
-    }
-
-    /// Returns the decoded message code.
-    #[must_use]
-    #[inline]
-    pub const fn code(self) -> MessageCode {
-        self.code
-    }
-
-    /// Returns the payload length encoded in the header.
-    #[must_use]
-    #[inline]
-    pub const fn payload_len(self) -> u32 {
-        self.payload_len
-    }
-
-    /// Returns the payload length encoded in the header as a native pointer-sized value.
-    ///
-    /// Upstream rsync's multiplexed payload field is limited to 24 bits, which comfortably fits
-    /// into the pointer widths supported by this project (32-bit and 64-bit architectures). The
-    /// conversion is therefore infallible, but a debug assertion is kept to surface potential
-    /// regressions if a narrower architecture is ever introduced.
-    #[must_use]
-    #[inline]
-    pub fn payload_len_usize(self) -> usize {
-        #[allow(clippy::assertions_on_constants)]
-        {
-            debug_assert!(
-                usize::BITS >= 24,
-                "multiplexed payloads require pointer widths of at least 24 bits"
-            );
-        }
-        self.payload_len as usize
-    }
-}
-
-impl TryFrom<[u8; HEADER_LEN]> for MessageHeader {
-    type Error = EnvelopeError;
-
-    #[inline]
-    fn try_from(bytes: [u8; HEADER_LEN]) -> Result<Self, Self::Error> {
-        Self::decode(&bytes)
-    }
-}
-
-impl TryFrom<&[u8; HEADER_LEN]> for MessageHeader {
-    type Error = EnvelopeError;
-
-    #[inline]
-    fn try_from(bytes: &[u8; HEADER_LEN]) -> Result<Self, Self::Error> {
-        Self::decode(bytes)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use proptest::prelude::*;
-
-    #[test]
-    fn header_round_trips_for_info_message() {
-        let header = MessageHeader::new(MessageCode::Info, 123).expect("constructible header");
-        let encoded = header.encode();
-        let decoded = MessageHeader::decode(&encoded).expect("decode succeeds");
-        assert_eq!(decoded, header);
-    }
-
-    #[test]
-    fn message_header_try_from_array_round_trips() {
-        let header = MessageHeader::new(MessageCode::Data, 17).expect("constructible header");
-        let encoded = header.encode();
-        let decoded = MessageHeader::try_from(encoded).expect("array conversion succeeds");
-
-        assert_eq!(decoded, header);
-    }
-
-    #[test]
-    fn message_header_try_from_slice_round_trips() {
-        let header = MessageHeader::new(MessageCode::Info, 7).expect("constructible header");
-        let encoded = header.encode();
-        let decoded = MessageHeader::try_from(&encoded).expect("slice conversion succeeds");
-
-        assert_eq!(decoded, header);
-    }
-
-    #[test]
-    fn message_header_try_from_array_rejects_invalid_tag() {
-        let encoded = [0u8; HEADER_LEN];
-        let err = MessageHeader::try_from(encoded).expect_err("invalid tag must fail");
-
-        assert_eq!(err, EnvelopeError::InvalidTag(0));
-    }
-
-    #[test]
-    fn message_header_new_supports_const_contexts() {
-        const HEADER: MessageHeader = match MessageHeader::new(MessageCode::Info, 42) {
-            Ok(header) => header,
-            Err(_) => panic!("valid header should be constructible in const context"),
-        };
-
-        assert_eq!(HEADER.code(), MessageCode::Info);
-        assert_eq!(HEADER.payload_len(), 42);
-    }
-
-    #[test]
-    fn message_header_from_raw_round_trips() {
-        let header =
-            MessageHeader::new(MessageCode::Stats, 0x0055_AA11).expect("constructible header");
-        let raw = header.encode_raw();
-        let decoded = MessageHeader::from_raw(raw).expect("raw representation decodes");
-
-        assert_eq!(decoded, header);
-    }
-
-    #[test]
-    fn message_header_from_raw_rejects_invalid_tag() {
-        let raw = 0x0000_0001u32; // tag without MPLEX_BASE offset
-        let err = MessageHeader::from_raw(raw).expect_err("invalid tag must fail");
-
-        assert_eq!(err, EnvelopeError::InvalidTag(0));
-    }
-
-    #[test]
-    fn message_header_from_raw_rejects_unknown_code() {
-        let raw = ((u32::from(MPLEX_BASE) + 0x40) << 24) | 0x0000_00FF;
-        let err = MessageHeader::from_raw(raw).expect_err("unknown code must fail");
-
-        assert_eq!(err, EnvelopeError::UnknownMessageCode(0x40));
-    }
-
-    #[test]
-    fn message_header_encode_raw_matches_encode() {
-        let header = MessageHeader::new(MessageCode::Success, 7).expect("constructible header");
-        let encoded = header.encode();
-        let raw = header.encode_raw();
-
-        assert_eq!(encoded, raw.to_le_bytes());
-    }
-
-    #[test]
-    fn try_from_log_code_converts_logging_variants() {
-        for &log in LogCode::all() {
-            match log {
-                LogCode::None => {
-                    let err = MessageCode::try_from(log).expect_err("FNONE has no multiplexed tag");
-                    assert_eq!(
-                        err,
-                        LogCodeConversionError::NoMessageEquivalent(LogCode::None)
-                    );
-                }
-                _ => {
-                    let message = MessageCode::try_from(log).expect("log code has multiplexed tag");
-                    assert_eq!(message.log_code(), Some(log));
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn try_from_message_code_rejects_non_logging_variants() {
-        for &code in MessageCode::all() {
-            match code.log_code() {
-                Some(log) => {
-                    let parsed = LogCode::try_from(code).expect("logging code maps to log code");
-                    assert_eq!(parsed, log);
-                }
-                None => {
-                    let err =
-                        LogCode::try_from(code).expect_err("non-logging message lacks log code");
-                    assert_eq!(err, LogCodeConversionError::NoLogEquivalent(code));
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn encode_into_slice_writes_bytes_without_touching_tail() {
-        let header =
-            MessageHeader::new(MessageCode::Info, 0x0055_AA11).expect("payload fits within header");
-        let mut buffer = [0xFFu8; HEADER_LEN + 3];
-
-        header
-            .encode_into_slice(&mut buffer)
-            .expect("buffer large enough for header");
-
-        assert_eq!(&buffer[..HEADER_LEN], &header.encode());
-        assert_eq!(buffer[HEADER_LEN..], [0xFFu8; 3]);
-    }
-
-    #[test]
-    fn encode_into_slice_rejects_short_buffers() {
-        let header = MessageHeader::new(MessageCode::Info, 7).expect("valid header");
-        let mut buffer = [0u8; HEADER_LEN - 1];
-
-        let err = header
-            .encode_into_slice(&mut buffer)
-            .expect_err("insufficient buffer should error");
-
-        assert_eq!(
-            err,
-            EnvelopeError::TruncatedHeader {
-                actual: HEADER_LEN - 1
-            }
-        );
-    }
-
-    #[test]
-    fn payload_len_usize_matches_u32_accessor() {
-        if usize::BITS < 24 {
-            // Architectures with pointer widths below 24 bits cannot represent the full
-            // multiplexed payload range. The implementation itself guards this via a debug
-            // assertion, so skip the comparison in that niche configuration.
-            return;
-        }
-
-        let header =
-            MessageHeader::new(MessageCode::Data, MAX_PAYLOAD_LENGTH).expect("max payload");
-        assert_eq!(header.payload_len_usize(), header.payload_len() as usize);
-    }
-
-    #[test]
-    fn message_header_encode_supports_const_contexts() {
-        const HEADER: MessageHeader = match MessageHeader::new(MessageCode::Warning, 7) {
-            Ok(header) => header,
-            Err(_) => panic!("valid header should be constructible in const context"),
-        };
-        const ENCODED: [u8; HEADER_LEN] = HEADER.encode();
-
-        assert_eq!(ENCODED, HEADER.encode());
-    }
-
-    #[test]
-    fn log_codes_are_hashable() {
-        use std::collections::HashSet;
-
-        let mut set = HashSet::new();
-        assert!(set.insert(LogCode::Info));
-        assert!(set.contains(&LogCode::Info));
-        assert!(!set.insert(LogCode::Info));
-    }
-
-    #[test]
-    fn message_codes_are_hashable() {
-        use std::collections::HashSet;
-
-        let mut set = HashSet::new();
-        assert!(set.insert(MessageCode::Data));
-        assert!(set.contains(&MessageCode::Data));
-        assert!(!set.insert(MessageCode::Data));
-    }
-
-    #[test]
-    fn decode_rejects_truncated_header() {
-        let err = MessageHeader::decode(&[0u8; 2]).unwrap_err();
-        assert_eq!(err, EnvelopeError::TruncatedHeader { actual: 2 });
-    }
-
-    #[test]
-    fn decode_rejects_tag_without_base_offset() {
-        let raw = (u32::from(MPLEX_BASE - 1) << 24) | 1;
-        let err = MessageHeader::decode(&raw.to_le_bytes()).unwrap_err();
-        assert_eq!(err, EnvelopeError::InvalidTag(MPLEX_BASE - 1));
-    }
-
-    #[test]
-    fn decode_rejects_unknown_message_codes() {
-        let unknown_code = 11u8;
-        let tag = u32::from(MPLEX_BASE) + u32::from(unknown_code);
-        let raw = (tag << 24) | 5;
-        let err = MessageHeader::decode(&raw.to_le_bytes()).unwrap_err();
-        assert_eq!(err, EnvelopeError::UnknownMessageCode(unknown_code));
-    }
-
-    #[test]
-    fn encode_uses_little_endian_layout() {
-        let payload_len = 0x00A1_B2C3;
-        let header =
-            MessageHeader::new(MessageCode::Info, payload_len).expect("constructible header");
-        let encoded = header.encode();
-
-        let expected_raw =
-            ((u32::from(MPLEX_BASE) + u32::from(MessageCode::Info.as_u8())) << 24) | payload_len;
-        assert_eq!(encoded, expected_raw.to_le_bytes());
-    }
-
-    #[test]
-    fn decode_masks_payload_length_to_24_bits() {
-        let tag = (u32::from(MPLEX_BASE) + u32::from(MessageCode::Info.as_u8())) << 24;
-        let raw = tag | (MAX_PAYLOAD_LENGTH + 1);
-        let header =
-            MessageHeader::decode(&raw.to_le_bytes()).expect("payload length is masked to 24 bits");
-        assert_eq!(header.code(), MessageCode::Info);
-        assert_eq!(
-            header.payload_len(),
-            (MAX_PAYLOAD_LENGTH + 1) & PAYLOAD_MASK
-        );
-    }
-
-    #[test]
-    fn new_rejects_oversized_payloads() {
-        let err = MessageHeader::new(MessageCode::Info, MAX_PAYLOAD_LENGTH + 1).unwrap_err();
-        assert_eq!(err, EnvelopeError::OversizedPayload(MAX_PAYLOAD_LENGTH + 1));
-    }
-
-    #[test]
-    fn message_code_variants_round_trip_through_try_from() {
-        for &code in MessageCode::all() {
-            let raw = code.as_u8();
-            let decoded = MessageCode::try_from(raw).expect("known code");
-            assert_eq!(decoded, code);
-        }
-    }
-
-    #[test]
-    fn message_code_into_u8_matches_as_u8() {
-        for &code in MessageCode::all() {
-            let converted: u8 = code.into();
-            assert_eq!(converted, code.as_u8());
-        }
-    }
-
-    #[test]
-    fn message_code_from_u8_matches_try_from() {
-        for &code in MessageCode::all() {
-            let raw = code.as_u8();
-            assert_eq!(MessageCode::from_u8(raw), Some(code));
-            assert_eq!(MessageCode::try_from(raw).ok(), MessageCode::from_u8(raw));
-        }
-    }
-
-    #[test]
-    fn message_code_from_u8_rejects_unknown_values() {
-        assert_eq!(MessageCode::from_u8(11), None);
-        assert_eq!(MessageCode::from_u8(0xFF), None);
-    }
-
-    #[test]
-    fn message_code_from_str_parses_known_names() {
-        for &code in MessageCode::all() {
-            let parsed: MessageCode = code.name().parse().expect("known name");
-            assert_eq!(parsed, code);
-        }
-    }
-
-    #[test]
-    fn message_code_from_str_rejects_unknown_names() {
-        let err = "MSG_SOMETHING_ELSE".parse::<MessageCode>().unwrap_err();
-        assert_eq!(err.invalid_name(), "MSG_SOMETHING_ELSE");
-        assert_eq!(
-            err.to_string(),
-            "unknown multiplexed message code name: \"MSG_SOMETHING_ELSE\""
-        );
-    }
-
-    #[test]
-    fn message_code_all_is_sorted_by_numeric_value() {
-        let all = MessageCode::all();
-        for window in all.windows(2) {
-            let first = window[0];
-            let second = window[1];
-            assert!(
-                first.as_u8() <= second.as_u8(),
-                "MessageCode::all() is not sorted: {:?}",
-                all
-            );
-        }
-    }
-
-    #[test]
-    fn header_round_trips_for_all_codes_and_sample_lengths() {
-        const PAYLOAD_SAMPLES: [u32; 3] = [0, 1, MAX_PAYLOAD_LENGTH];
-
-        for &code in MessageCode::all() {
-            for &len in &PAYLOAD_SAMPLES {
-                let header = MessageHeader::new(code, len).expect("constructible header");
-                let encoded = header.encode();
-                let decoded = MessageHeader::decode(&encoded).expect("decode succeeds");
-                assert_eq!(decoded.code(), code);
-                assert_eq!(decoded.payload_len(), len);
-            }
-        }
-    }
-
-    #[test]
-    fn logging_classification_matches_upstream_set() {
-        const LOGGING_CODES: &[MessageCode] = &[
-            MessageCode::ErrorXfer,
-            MessageCode::Info,
-            MessageCode::Error,
-            MessageCode::Warning,
-            MessageCode::ErrorSocket,
-            MessageCode::ErrorUtf8,
-            MessageCode::Log,
-            MessageCode::Client,
-        ];
-
-        for &code in MessageCode::all() {
-            let expected = LOGGING_CODES.contains(&code);
-            assert_eq!(code.is_logging(), expected, "mismatch for code {code:?}");
-        }
-    }
-
-    #[test]
-    fn message_code_name_matches_upstream_identifiers() {
-        use MessageCode::*;
-
-        let expected = [
-            (Data, "MSG_DATA"),
-            (ErrorXfer, "MSG_ERROR_XFER"),
-            (Info, "MSG_INFO"),
-            (Error, "MSG_ERROR"),
-            (Warning, "MSG_WARNING"),
-            (ErrorSocket, "MSG_ERROR_SOCKET"),
-            (Log, "MSG_LOG"),
-            (Client, "MSG_CLIENT"),
-            (ErrorUtf8, "MSG_ERROR_UTF8"),
-            (Redo, "MSG_REDO"),
-            (Stats, "MSG_STATS"),
-            (IoError, "MSG_IO_ERROR"),
-            (IoTimeout, "MSG_IO_TIMEOUT"),
-            (NoOp, "MSG_NOOP"),
-            (ErrorExit, "MSG_ERROR_EXIT"),
-            (Success, "MSG_SUCCESS"),
-            (Deleted, "MSG_DELETED"),
-            (NoSend, "MSG_NO_SEND"),
-        ];
-
-        for &(code, name) in &expected {
-            assert_eq!(code.name(), name);
-            assert_eq!(code.to_string(), name);
-        }
-    }
-
-    #[test]
-    fn message_code_flush_alias_matches_info() {
-        assert_eq!(MessageCode::FLUSH, MessageCode::Info);
-        assert_eq!(MessageCode::FLUSH.as_u8(), MessageCode::Info.as_u8());
-
-        let parsed: MessageCode = "MSG_FLUSH".parse().expect("known alias");
-        assert_eq!(parsed, MessageCode::Info);
-    }
-
-    proptest! {
-        #[test]
-        fn prop_header_round_trips_for_random_inputs(
-            code in proptest::sample::select(&MessageCode::ALL),
-            payload in 0u32..=MAX_PAYLOAD_LENGTH,
-        ) {
-            let header = MessageHeader::new(code, payload).expect("constructible header");
-            let encoded = header.encode();
-            let decoded = MessageHeader::decode(&encoded).expect("decode succeeds");
-            prop_assert_eq!(decoded, header);
-        }
-
-        #[test]
-        fn prop_decode_rejects_all_truncated_header_lengths(len in 0usize..HEADER_LEN) {
-            let bytes = vec![0u8; len];
-            let err = MessageHeader::decode(&bytes).unwrap_err();
-            prop_assert_eq!(err, EnvelopeError::TruncatedHeader { actual: len });
-        }
-
-        #[test]
-        fn prop_decode_rejects_invalid_tag_prefixes(
-            tag in 0u8..MPLEX_BASE,
-            payload in 0u32..=PAYLOAD_MASK,
-        ) {
-            let raw = (u32::from(tag) << 24) | (payload & PAYLOAD_MASK);
-            let err = MessageHeader::decode(&raw.to_le_bytes()).unwrap_err();
-            prop_assert_eq!(err, EnvelopeError::InvalidTag(tag));
-        }
-
-        #[test]
-        fn prop_message_header_new_rejects_oversized_payloads(
-            len in (MAX_PAYLOAD_LENGTH + 1)..=(MAX_PAYLOAD_LENGTH + 0x0FFF),
-        ) {
-            let err = MessageHeader::new(MessageCode::Info, len).unwrap_err();
-            prop_assert_eq!(err, EnvelopeError::OversizedPayload(len));
-        }
-
-        #[test]
-        fn prop_decode_rejects_unknown_message_codes(
-            code in 0u8..=(u8::MAX - MPLEX_BASE),
-            payload in 0u32..=PAYLOAD_MASK,
-        ) {
-            prop_assume!(MessageCode::from_u8(code).is_none());
-            let tag = u32::from(MPLEX_BASE) + u32::from(code);
-            let raw = (tag << 24) | (payload & PAYLOAD_MASK);
-            let err = MessageHeader::decode(&raw.to_le_bytes()).unwrap_err();
-            prop_assert_eq!(err, EnvelopeError::UnknownMessageCode(code));
-        }
-    }
-
-    #[test]
-    fn log_code_all_is_sorted_by_numeric_value() {
-        let all = LogCode::all();
-        for window in all.windows(2) {
-            let first = window[0];
-            let second = window[1];
-            assert!(
-                first.as_u8() <= second.as_u8(),
-                "LogCode::all() unsorted: {all:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn log_code_from_u8_matches_try_from() {
-        for &code in LogCode::all() {
-            let raw = code.as_u8();
-            assert_eq!(LogCode::from_u8(raw), Some(code));
-            assert_eq!(LogCode::try_from(raw).ok(), LogCode::from_u8(raw));
-        }
-    }
-
-    #[test]
-    fn log_code_from_u8_rejects_unknown_values() {
-        assert_eq!(LogCode::from_u8(9), None);
-        let err = LogCode::try_from(9).unwrap_err();
-        assert_eq!(err.invalid_value(), Some(9));
-        assert_eq!(err.to_string(), "unknown log code value: 9");
-    }
-
-    #[test]
-    fn log_code_from_str_parses_known_names() {
-        for &code in LogCode::all() {
-            let parsed: LogCode = code.name().parse().expect("known log code name");
-            assert_eq!(parsed, code);
-        }
-    }
-
-    #[test]
-    fn log_code_from_str_rejects_unknown_names() {
-        let err = "FUNKNOWN".parse::<LogCode>().unwrap_err();
-        assert_eq!(err.invalid_name(), Some("FUNKNOWN"));
-        assert_eq!(err.to_string(), "unknown log code name: \"FUNKNOWN\"");
-        assert_eq!(err.invalid_value(), None);
-    }
-
-    #[test]
-    fn log_code_name_matches_upstream_identifiers() {
-        use LogCode::*;
-
-        let expected = [
-            (None, "FNONE"),
-            (ErrorXfer, "FERROR_XFER"),
-            (Info, "FINFO"),
-            (Error, "FERROR"),
-            (Warning, "FWARNING"),
-            (ErrorSocket, "FERROR_SOCKET"),
-            (Log, "FLOG"),
-            (Client, "FCLIENT"),
-            (ErrorUtf8, "FERROR_UTF8"),
-        ];
-
-        for &(code, name) in &expected {
-            assert_eq!(code.name(), name);
-            assert_eq!(code.to_string(), name);
-        }
-    }
-
-    #[test]
-    fn message_code_log_code_matches_logging_subset() {
-        for &code in MessageCode::all() {
-            let log_code = code.log_code();
-            assert_eq!(
-                log_code.is_some(),
-                code.is_logging(),
-                "mismatch for {code:?}"
-            );
-
-            if let Some(mapped) = log_code {
-                assert!(matches!(
-                    mapped,
-                    LogCode::ErrorXfer
-                        | LogCode::Info
-                        | LogCode::Error
-                        | LogCode::Warning
-                        | LogCode::ErrorSocket
-                        | LogCode::Log
-                        | LogCode::Client
-                        | LogCode::ErrorUtf8
-                ));
-            }
-        }
-    }
-
-    #[test]
-    fn message_code_from_log_code_round_trips_logging_variants() {
-        for &log in LogCode::all() {
-            match MessageCode::from_log_code(log) {
-                Some(code) => {
-                    assert_eq!(code.log_code(), Some(log), "round-trip failed for {log:?}");
-                }
-                None => assert_eq!(log, LogCode::None, "only FNONE lacks a multiplexed tag"),
-            }
-        }
-    }
-
-    #[test]
-    fn message_code_from_log_code_rejects_none_variant() {
-        assert_eq!(MessageCode::from_log_code(LogCode::None), None);
-    }
-
-    #[test]
-    fn try_from_log_code_maps_logging_variants() {
-        for &log in LogCode::all() {
-            match MessageCode::try_from(log) {
-                Ok(code) => assert_eq!(code.log_code(), Some(log)),
-                Err(err) => {
-                    assert_eq!(log, LogCode::None);
-                    assert_eq!(err.log_code(), Some(log));
-                    assert!(err.message_code().is_none());
-                    assert_eq!(
-                        err.to_string(),
-                        "log code FNONE has no multiplexed message equivalent"
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn try_from_message_code_requires_logging_equivalent() {
-        for &code in MessageCode::all() {
-            match LogCode::try_from(code) {
-                Ok(log) => assert_eq!(MessageCode::from_log_code(log), Some(code)),
-                Err(err) => {
-                    assert!(code.log_code().is_none());
-                    assert_eq!(err.message_code(), Some(code));
-                    assert!(err.log_code().is_none());
-                    assert_eq!(
-                        err.to_string(),
-                        format!("message code {code} has no log code equivalent")
-                    );
-                }
-            }
-        }
-    }
-}
